add RESTful API of user managerment

+++ src/routes/category/category.controller.js
@@ -1,14 +1,16 @@
 const category_service = require('../../models/Category/category.service');
 const subCategory_service = require('../../models/SubCategory/subCategory.service');
-// const jwt_helper = require('../../helpers/jwt.helper');
+const { getTotalPage } = require('../../helpers/etc.helper');
 const http_message = require('../../constants/http_message.constant');
 
-module.exports.getCategory = async (req, res) => {
-	const all_category = await category_service.getAll([]);
+module.exports.getNestedCategory = async (req, res) => {
+	const list = await category_service.getAll([]);
+	const all_category = list.rows;
 	const rs = { count: all_category.length, data: [] };
 
 	for (const item of all_category) {
-		const all_sub_category = await subCategory_service.getAllByCategoryId(item.category_id, [ 'category_id' ]);
+		const list = await subCategory_service.getAllByCategoryId(item.category_id, []);
+		const all_sub_category = list.rows;
 		const item_category = {
 			category_id: item.category_id,
 			name: item.name,
@@ -21,6 +23,18 @@ module.exports.getCategory = async (req, res) => {
 		}
 		rs.data.push(item_category);
 	}
+
+	return res.json(rs);
+};
+
+module.exports.getCategory = async (req, res) => {
+	const { page, limit } = req.query;
+	const rs = {};
+	const list = await category_service.getAll([], page, limit);
+	rs.count = list.count;
+	rs.data = list.rows;
+	rs.page = +page;
+	rs.total_page = getTotalPage(rs.count, limit);
 
 	return res.json(rs);
 };
@@ -41,7 +55,12 @@ module.exports.updateCategoryName = async (req, res) => {
 
 module.exports.deleteCategory = async (req, res) => {
 	const category_id = req.params.id;
-	// kiểm tra category có danh mục con ko, có thì xóa luôn
+	const list = await subCategory_service.getAllByCategoryId(category_id, [], 1, 99999999999);
+
+	if (list.rows.length > 0) {
+		return res.status(400).json({ errs: [ http_message.status400_exist_sub_category.message ] });
+	}
+
 	await category_service.deleteCategory(category_id);
 
 	return res.json(http_message.status200);
