+++ src/routes/subCategory/subCategory.controller.js
@@ -1,8 +1,18 @@
 const subCategory_service = require('../../models/SubCategory/subCategory.service');
 const http_message = require('../../constants/http_message.constant');
+const product_service = require('../../models/Product/Product.service');
+const { getTotalPage } = require('../../helpers/etc.helper');
 
 module.exports.getByCategory = async (req, res) => {
-	return res.json('ok');
+	const { limit, page, category_id } = req.query;
+	const rs = {};
+	const list = await subCategory_service.getAllByCategoryId(category_id, [], page, limit);
+	rs.count = list.count;
+	rs.data = list.rows;
+	rs.page = +page;
+	rs.total_page = getTotalPage(rs.count, limit);
+
+	return res.json(rs);
 };
 
 module.exports.createSubCategoryPost = async (req, res) => {
@@ -15,18 +25,24 @@ module.exports.createSubCategoryPost = async (req, res) => {
 module.exports.updateSubCategory = async (req, res) => {
 	const { sub_category_id, name, category_id } = req.body;
 
-  if(name) {
-    await subCategory_service.updateName(name, sub_category_id);
-  }
-  if(category_id) {
-    await subCategory_service.updateCategoryId(category_id, sub_category_id);
-  }
+	if (name) {
+		await subCategory_service.updateName(name, sub_category_id);
+	}
+	if (category_id) {
+		await subCategory_service.updateCategoryId(category_id, sub_category_id);
+	}
 
 	return res.json(http_message.status200);
 };
 
 module.exports.deleteSubCategory = async (req, res) => {
 	const sub_category_id = req.params.id;
+	const list = await product_service.getManyProductBySubCategory(sub_category_id, []);
+
+	if (list.length > 0) {
+		return res.status(400).json({ errs: [ http_message.status400_exist_product.message ] });
+	}
+
 	await subCategory_service.deleteSubCategory(sub_category_id);
 
 	return res.json(http_message.status200);
